+++ packages/localize-tools/src/tests/e2e/build-transform-xliff.test.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-import {e2eGoldensTest} from './e2e-goldens-test';
+import {e2eGoldensTest} from './e2e-goldens-test.js';
 
 e2eGoldensTest('build-transform-xliff', [
   '--config=lit-localize.json',