+++ packages/localize-tools/src/tests/e2e/extract-xlb.test.ts
@@ -4,6 +4,6 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-import {e2eGoldensTest} from './e2e-goldens-test';
+import {e2eGoldensTest} from './e2e-goldens-test.js';
 
 e2eGoldensTest('extract-xlb', ['--config=lit-localize.json', 'extract']);